Introduction of AppCompositionRoot
