+++ app/src/main/java/com/techyourchance/dagger2course/MyApplication.kt
@@ -1,26 +1,14 @@
 package com.techyourchance.dagger2course
 
 import android.app.Application
-import com.techyourchance.dagger2course.networking.StackoverflowApi
-import com.techyourchance.dagger2course.questions.FetchQuestionDetailsUseCase
-import com.techyourchance.dagger2course.questions.FetchQuestionsUseCase
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
+import com.techyourchance.dagger2course.common.composition.AppCompositionRoot
 
 class MyApplication: Application() {
 
-    private val retrofit: Retrofit = Retrofit.Builder()
-            .baseUrl(Constants.BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-
-    private val stackoverflowApi: StackoverflowApi = retrofit.create(StackoverflowApi::class.java)
-
-    public val fetchQuestionsUseCase get() = FetchQuestionsUseCase(stackoverflowApi)
-
-    public val fetchQuestionDetailsUseCase get() = FetchQuestionDetailsUseCase(stackoverflowApi)
+    public lateinit var appCompositionRoot: AppCompositionRoot
 
     override fun onCreate() {
+        appCompositionRoot = AppCompositionRoot()
         super.onCreate()
     }
